Swap wiki sign up to use google ReCaptchaNoCaptcha
Swap to the more modern google No CAPTCHA reCAPTCHA away from the older
google reCAPTCHA which gathers address data for google maps.

This simple config swap assumes that the recent ConfirmEdit version we
have installed (1.4) comes with all the necessary bits to enable this
captcha class and that Google's old ReCaptcha uses the same API keys as
it's new one.

In future we might figure out a solution which also doesn't request
any javascript from google but really, not contributing to google maps
with every OSM wiki sign up would be a step in the right direction

Fixes: https://github.com/openstreetmap/operations/issues/19