Output TagInfo <script> tag
For the "[taginfo taglists](https://wiki.openstreetmap.org/wiki/Taginfo/Taglists)" on the wiki. This thing is already in place and working, but so far we've kept a chunk of code on https://wiki.openstreetmap.org/wiki/MediaWiki:Common.js and Jochen Topf has been asking wiki admins to make little tweaks to it. He wanted to host it at his end instead.

The approach for having mediawiki spit out a <script> tag is [based on this answer](http://stackoverflow.com/a/26436873/338265).