delete the planet.db sqlite file on each run
at least as a temporary solution to ensure spam gets blown away: https://github.com/gravitystorm/blogs.osm.org/issues/17

The sqlite file gets rebuilt completely on pluto build. Doing this on each iteration may not be what the pluto developer intended, but the result is the same ...but without the spam!