Use multipackage installs throughout
In Chef 12.1 we introduced multipackage installs. Instead of shelling
out to dpkg for each package to determine what we should do we gather
the information we need and check each package in the array. It's faster
and it takes less memory. It's especially useful after the first run when
nothing actually changes. Far fewer resources consumed.

I noticed you arranged the packages so I tried to keep them to that
same arrangement and I used multiline arrays since they make the diffs
nicer in the future as you add / remove packages. This isn't necessary
for Chef 13, but it's best practice.

Signed-off-by: Tim Smith <tsmith@chef.io>

Closes #137