Remove RRD dumps from Munin server
In #109 we added downloadable RRD dumps, which may or may not have been useful at the time. However, they've been broken for almost a year (since August 2018 by my reckoning) without anyone complaining. Except `cron`, which complains to me daily.

If, as it appears, it's not being used any more, then it seems we might as well turn it off.