Allow retrying of apache service actions
There seems to be an issue with the otrs cookbook when deployed
on a new machine where it tries to both restart and also reload
apache and the reload sometime fails because it seems to happen
before the restart has completed.

It's possible this is because the use of mod_perl for otrs is
slowing down the restart but as a workaround we allow the reload
action to be retried a couple of times.