Bump planet-dump-ng version for tag order change.
Update to the latest version of planet-dump-ng, which changes the order of tags in the output file to match that used by other tools. Planet-dump-ng already has a _consistent_ ordering, but not a natural one. The new ordering sorts tag keys as UTF-8 encoded strings of bytes, so it's not a "proper" Unicode string sort but still more understandable than the previous method.

See [osmdbt#29](https://github.com/openstreetmap/osmdbt/issues/29) for previous discussion and more information.