DMCA form: add Content-Type header to generated email
DMCA form mails lack a content-type header. Quite often non-latin, such as Persian, is filled in, typically in utf-8, which currently looks garbled in the receiving mail client.
This PR adds the same content-type that the html page uses itself, and is the most likely to be filled by the form users.