Decrease tile rate limit to 30 TPS
This is 15 TPS per server, except for pyrene which needs 30.

Now that the rate limiting is working I reviewed IPs exceeding 20
TPS on an hour basis, looking at several days, and they were all
scrapers, or fakes which I would normally block if not for the effort.
I found no non-scraping proxies or NAT.